Update Cache section with dos2unix conversions
The default to checkout files on windows converts the line endings to
windows format, however since the files are directly exposed to the linux
container they need to be in unix format to function correctly. dos2unix
does the correct conversion on the files to make sure the scripts work
as intended.